backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
sqlalchemy==2.0.23
python-dotenv==1.0.0
pyjwt==2.8.0
bcrypt==4.1.1
python-multipart==0.0.6
PyPDF2==3.0.1
pdfplumber==0.10.3
groq==0.4.2
aiosqlite==0.19.0
pageindex==0.1.5
litellm==1.83.0
pymupdf==1.26.4
pyyaml==6.0.2

# RAG dependencies (Python 3.12 compatible)
sentence-transformers>=3.0.0
numpy>=1.26.0
